Fixed the switches not working properly
This also meant a few reverts of vanilla collision.  I'll do this slowly from now on, it is a very delicate area.

+++ src/engine/p_mobj.h
@@ -199,7 +199,7 @@ typedef struct mobj_s {
 
     int			        tics;	// state tic counter
     state_t*            state;
-    intptr_t			flags;
+    int			        flags;
     int			        health;
 
     // [d64] alpha value for rendering
